extract anonymous class
